+++ CryptoLib/sha-1.h
@@ -35,4 +35,4 @@ typedef struct _Sha1State {
 #define HASH_STATE_SHA1_SIZE             HASH_STATE_HEADER_SIZE + sizeof(Sha1State)
 
 void Sha1InitState(__out uint32_t* state);
-void Sha1Get(__inout Sha1State* state, __in_opt const void* input, __in uint64_t inputSize, __in bool finalize, __out_opt uint32_t* output);
+void Sha1Get(__inout Sha1State* state, __in_opt const void* input, __in size_t inputSize, __in bool finalize, __out_opt uint32_t* output);
