+++ CryptoTests/ProcessingByBlockCipherDecryptionT.cpp
@@ -10,7 +10,7 @@
 TEST(ProcessingByBlockCipherDecryptionTest, TooSmallOutputSize) {
     int status = NO_ERROR;
     uint8_t input[] = { 0xb9, 0xe9, 0x8a, 0x3c, 0x77, 0xa5, 0x10, 0x86 };
-    uint64_t outputSize = 0;
+    size_t outputSize = 0;
     uint8_t* buffer = new uint8_t[outputSize];
     BlockCipherHandle handle = nullptr;
     EVAL(InitBlockCipherState(&handle, DES_cipher_type, Decryption_mode, ECB_mode, PKCSN7_padding, KEY_8, nullptr));
